units: make BPM a strong au type (QuarterNote / Minutes)
Replace the raw float bpm_t typedef with au::QuantityD<BPM> where
BPM = QuarterNote / au::Minutes, sharing the TickDim × TimeDim⁻¹
dimension with Tick/Second. This enables direct dimensional arithmetic
in tempo conversions (ticks / bpm × sample_rate → samples), eliminating
manual factor computation and / 60.0 conversions throughout the DSP
layer.

- Define BPM, bpm, bpm_t in units.h; remove old typedef from types.h
- Refactor TempoMap to use bpm_t for TempoEvent, add_tempo_event,
  tempo_at_tick, and all internal tick↔seconds conversions
- Simplify ContentTimeWarp and TempoWarpMap to use source_bpm_
  directly in dimensional arithmetic instead of extracting a scalar
  factor
- Simplify RegionRenderer and AudioRegion to use dimensional
  arithmetic (ticks / bpm × sample_rate) instead of manual
  factor = (bpm / 60) × ppq intermediates
- Change detect_bpm() and sourceBpm() return types from float to
  double for consistency with all other QuantityD types
- Update all call sites and tests across the codebase

+++ src/structure/arrangement/audio_region.h
@@ -25,7 +25,7 @@ class AudioRegion final
   Q_PROPERTY (
     AudioRegion::MusicalMode musicalMode READ musicalMode WRITE setMusicalMode
       NOTIFY musicalModeChanged)
-  Q_PROPERTY (float sourceBpm READ sourceBpm CONSTANT)
+  Q_PROPERTY (double sourceBpm READ sourceBpm CONSTANT)
   Q_PROPERTY (
     bool effectivelyInMusicalMode READ effectivelyInMusicalMode NOTIFY
       effectivelyInMusicalModeChanged)
@@ -69,7 +69,7 @@ public:
 
   MusicalMode   musicalMode () const { return musical_mode_; }
   bool          effectivelyInMusicalMode () const;
-  float         sourceBpm () const;
+  double        sourceBpm () const;
   void          setMusicalMode (MusicalMode musical_mode);
   Q_SIGNAL void musicalModeChanged (MusicalMode musical_mode);
   Q_SIGNAL void effectivelyInMusicalModeChanged ();
